feat: Layout.tsx 서버 컴포넌트 유지할 수 있도록 수정
