+++ src/components/layout/Header/Header.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import IconButton from "@/components/common/Button/IconButton";
 import {
   LogoLineIcon,
@@ -7,14 +9,22 @@ import {
 import Searchbar from "@/components/Searchbar/Searchbar";
 import { ArrowLeftIcon, LogIn, MenuIcon } from "lucide-react";
 import Link from "next/link";
+import { usePathname } from "next/navigation";
 import { useEffect, useState } from "react";
 
 interface HeaderProps {
   onMenuClick: () => void;
-  onLoginClick: () => void;
 }
 
-const Header = ({ onMenuClick, onLoginClick }: HeaderProps) => {
+const Header = ({ onMenuClick }: HeaderProps) => {
+  const pathname = usePathname();
+  const hideHeaderRoutes = [
+    "/login",
+    "/signup",
+    "/password-forget",
+    "/password-reset",
+    "/email-validation",
+  ];
   const [isMobileSearchOpen, setIsMobileSearchOpen] = useState(false);
 
   useEffect(() => {
@@ -31,6 +41,8 @@ const Header = ({ onMenuClick, onLoginClick }: HeaderProps) => {
       window.removeEventListener("resize", handleResize);
     };
   }, []);
+
+  if (hideHeaderRoutes.includes(pathname)) return null;
 
   return (
     <header className="h-[3rem] px-4 sm:px-10 flex items-center justify-between">
@@ -77,12 +89,9 @@ const Header = ({ onMenuClick, onLoginClick }: HeaderProps) => {
             />
           </div>
           <IconButton icon={<NotificationIcon />} isTransparent label="알림" />
-          <IconButton
-            icon={<LogIn />}
-            isTransparent
-            label="로그인"
-            onClick={onLoginClick}
-          />
+          <Link href="/login">
+            <IconButton icon={<LogIn />} isTransparent label="로그인" />
+          </Link>
         </div>
       )}
     </header>
