+++ src/components/layout/HeaderShell.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { useState } from "react";
+import Header from "./Header/Header";
+import Sidebar from "./Sidebar/Sidebar";
+
+export default function HeaderShell() {
+  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+
+  return (
+    <>
+      <Header onMenuClick={() => setIsSidebarOpen(true)} />
+      <Sidebar isOpen={isSidebarOpen} onClose={() => setIsSidebarOpen(false)} />
+    </>
+  );
+}
